feat: migrate icon component
