+++ src/icon/IconWrapper.styles.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import styled from 'styled-components';
 
 import { backgroundColors } from './theme';
