--- src/icon/index.js
@@ -1,3 +1,0 @@
-import Icon from './Icon';
-
-export default Icon;
