+++ src/icon/index.tsx
@@ -1,0 +1,5 @@
+import Icon, { IconProps } from './Icon';
+
+export type { IconProps };
+
+export default Icon;
